remove moviePage container and replace connect to moviePage component
