+++ src/app/App.jsx
@@ -4,7 +4,7 @@ import { createBrowserHistory } from 'history';
 
 import Header from '../components/Header';
 import { ROUTES, MOVIE_TYPE as type } from '../constants/constants';
-import MoviePage from '../containers/MoviePage';
+import MoviePage from '../page/MoviePage';
 
 import './app.css';
